+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.4 [unreleased]
 
+- Added `scan-object-acl` command
 - Added `--log-bucket` and `--log-prefix` options to `scan` command
 - Added `--log-prefix` option to `enable-logging` command